Move DistanceMatrix to services as it depends on other plugins
